backend/requirements/development.txt
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
pytest-asyncio>=0.23,<1.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0

# Linting & Formatting
ruff>=0.1,<1.0

# Type checking
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Dev tools
pre-commit>=3.6,<4.0
ipython>=8.0,<9.0
django-debug-toolbar>=4.2,<5.0
